apps/server/pyproject.toml
[project]
name = "pdf-master-server"
version = "0.1.0"
description = "PDF Master Backend Server"
authors = ["PDF Master Team"]
readme = "README.md"
requires-python = ">=3.10"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
aiofiles = "^23.2.0"

# PDF处理
pymupdf = "^1.23.0"
pikepdf = "^8.10.0"
pdfplumber = "^0.10.0"
reportlab = "^4.0.0"
pdf2docx = "^0.5.0"
python-docx = "^1.1.0"
openpyxl = "^3.1.0"
python-pptx = "^0.6.0"

# 图像处理
pillow = "^10.2.0"
pdf2image = "^1.17.0"

# OCR
paddleocr = "^2.7.0"
paddlepaddle = "^2.5.0"

# AI
openai = "^1.10.0"
langchain = "^0.1.0"

# 数据库
sqlalchemy = "^2.0.0"
aiosqlite = "^0.19.0"

# 缓存
redis = "^5.0.0"

# 任务队列
celery = "^5.3.0"

# 工具
httpx = "^0.26.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
black = "^24.1.0"
ruff = "^0.1.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
